requirements.txt
pyautogui>=0.9.54
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
PySide6>=6.5.0
pyinstaller>=6.0.0